thursday/requirements.txt
PyYAML==6.0.1
deepdiff==7.0.1
redis==4.3.3
pymongo==4.8.0
Flask==2.2.5
flatten-dict==0.4.2
asteval==0.9.23
MonthDelta==1.0b0
pyjson5==1.6.6
pyflakes==1.6.0
nexmo==2.4.0
plivo==4.3.2
smpplib==2.0.1
pytz==2019.3
pbkdf2==1.3
webassets==2.0
marshmallow==3.19.0
wachtwoord==0.14
minio==5.0.7
ipcalc==1.99.0
celery==5.5.3
celery-once==3.0.1
zope.event>=4.6

Flask-Assets==2.0
decorator==4.3.0
Werkzeug<3.0

validators==0.12.2
tzlocal==1.5.1
ujson==1.35
itsdangerous==2.0
requests==2.27.1

chroma==0.2.0